refactor: reduce constructors in LeaseFileUploadLock to 2 and streamline lease creation

+++ src/main/java/me/desair/tus/server/upload/disk/DiskStorageService.java
@@ -380,7 +380,9 @@ public class DiskStorageService extends AbstractDiskBasedService implements Uplo
         Files.deleteIfExists(checksumFile);
       }
       Path uploadPath = getPathInStorageDirectory(info.getId());
-      FileUtils.deleteDirectory(uploadPath.toFile());
+      if (uploadPath != null) {
+        FileUtils.deleteDirectory(uploadPath.toFile());
+      }
     }
   }
